refactor: extract treeshake var and add LSP code actions, completions, d
- Move treeshakeEnabled check to avoid mutating file.Config
- Extract walkToPrimary helper in null-check detection
- Add LSP code_actions.go with import-insertion quick fixes
- Add LSP completions, definition, hover, and signature_help files
